Fix event emitter memory leak
NOTE: We use useEffect in each child PangPonent that depends on the
results of the PangContext initialization. This boilerplate may be
reduced if we find an elegant way to share common references to each
individual component state in the common parent component. For now we
can manage with the boilerplate.